+++ packages/settings-selector/src/ModeIcon/index.js
