+++ packages/settings-selector/src/ModeIcon/styled.js
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+const ModeIconWrap = styled.div`
+  width: 100px;
+  svg {
+    width: 100%;
+  }
+`;
+
+export default ModeIconWrap;
