+++ packages/settings-selector/src/SubmodeSelector/styled.js
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+export const SubmodeRow = styled.div`
+  > * {
+    width: inherit;
+    padding: 3px 0px 3px 5px;
+  }
+  button {
+    padding: 6px 12px;
+  }
+  svg,
+  img {
+    margin-left: 0px;
+  }
+`;
+
+export default SubmodeRow;
